requirements.txt
Flask
firebase-admin
torch
torchvision
opencv-python-headless
Pillow
numpy
pandas
scipy